chore: move cards to components folder
