+++ client/hosting/server-settings/components/phpmyadmin-card/index.js
