+++ client/hosting/server-settings/components/phpmyadmin-card/restore-db-password.js
