+++ client/hosting/server-settings/components/phpmyadmin-card/style.scss
