+++ client/hosting/server-settings/components/restore-plan-software-card/index.js
