+++ client/hosting/server-settings/components/sftp-card/sftp-card-loading-placeholder.tsx
