+++ client/hosting/server-settings/components/sftp-card/ssh-keys.scss
