+++ client/hosting/server-settings/components/sftp-card/use-detach-ssh-key.ts
